More refactoring, Put functionality back into Render
